Update service

See merge request !7

+++ src/main/java/com/epam/labproject/service/UserService.java
@@ -30,23 +30,32 @@ public class UserService {
     }
 
     public void save(User user){
-        user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
-        user.setRole(roleService.findByName("USER"));
-        userRepository.save(user);
-        PasswordEncoderFactories.createDelegatingPasswordEncoder();
+        if(user!=null) {
+            if(userRepository.findByLogin(user.getLogin())==null) {
+                user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+                user.setRole(roleService.findByName("USER"));
+                userRepository.save(user);
+                PasswordEncoderFactories.createDelegatingPasswordEncoder();
+            }else{
+                System.out.println("already exist page have to be here");
+            }
+        }
     }
     public User delete(String login){
         User user = userRepository.findByLogin(login);
+        if(user!=null)
         userRepository.delete(user);
         return user;
     }
 
-    public User update(User user){
-        if(userRepository.findByLogin(user.getLogin())==null){
-            this.save(user);
-            return null;
+    public void update(User user){
+        if(user!=null){
+            User currentUser=userRepository.findByLogin(user.getLogin());
+            if(currentUser!=null){
+                currentUser.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+                currentUser.setLogin(user.getLogin());
+                userRepository.save(currentUser);
+            }
         }
-        //userRepository.
-        return null;
     }
 }
